acu170189 - enhance cloudstack ohai plugin to read metadata from dhcp lease server
