--- spec/chef/ohai/mixin/test_meta-data_dir/public-hostname
@@ -1,1 +1,0 @@
-my_public_host_name
