--- spec/chef/ohai/mixin/test_meta-data_dir/public-ip
@@ -1,1 +1,0 @@
-1.2.3.4
