--- spec/chef/ohai/mixin/test_meta-data_dir/sub_hash/foo
@@ -1,1 +1,0 @@
-bar
